refactored some of the code to work together

+++ app/src/main/java/edu/ucsd/cse110/cse110group8_compass/Location.java
@@ -46,8 +46,8 @@ public class Location extends AppCompatActivity {
             String lat = latitude.getText().toString();
             String longit = longitude.getText().toString();
             Pin defaultPin = new Pin();
-            Pin parent = new Pin("Testparent",Integer.parseInt(lat),
-                    Integer.parseInt(longit));
+            Pin parent = new Pin("Testparent",Double.parseDouble(lat),
+                    Double.parseDouble(longit));
             intent.putExtra("label", parent.name);
             intent.putExtra("latitude", lat);
             intent.putExtra("longitude", longit);
